+++ app/scripts/components/landing-page/main.jsx
