--- app/scripts/components/test.jsx
@@ -1,13 +1,0 @@
-var React = require('react'); 
-
-var MyTest = React.createClass({
-	render: function() {
-		return (
-			<div>
-				<p>This is Gerty doing what she does best Which is ???</p>
-			</div>
-		);
-	}
-});
-
-module.exports = MyTest;
