+++ app/scripts/components/tests/content.jsx
@@ -1,0 +1,15 @@
+(function() {
+	var React = require('react');
+
+		var Content = React.createClass({
+			render: function() {
+				return (
+					<div>
+						<p>This is some content. blase blase</p>
+					</div>
+				);
+			}
+		});
+
+		module.exports = Content;
+})();
