+++ package.json
@@ -31,14 +31,14 @@
     "jade": "~1.11.0",
     "jshint": "~2.8.0",
     "karma": "^0.13.15",
-    "material-ui": "^0.13.2",
     "morgan": "~1.6.1",
     "path": "~0.12.7",
     "pg": "^4.4.3",
     "pg-hstore": "^2.3.2",
-    "react": "^0.14.2",
+    "react": "^0.13.3",
     "react-dom": "^0.14.2",
     "react-jade": "^2.5.0",
+    "react-router": "^0.13.3",
     "reactify": "^1.1.1",
     "sequelize": "~3.13.0",
     "serve-favicon": "~2.3.0",
